# Calculations optimized. Cache is compiled with the binary instead of being read at runtime.