- Perfect Hashing Functions (PHF)
- Realtime + Calculator optimizations
- Code simplification and removed functions with too many arguments
- Items to be simulated increased from 113 to 169
- Average `fn realtime` execution time reduced from 2.2ms to 0.4ms (82% reduction)
    - Size of JSON returned on average = enemy count * 193KB ~ 1MB on a real game
- Average `fn calculator` execution time reduced from 3ms to 0.2ms (93% reduction)
    - Size of JSON returned = enemy count * 129.85 KB
- Tests will have to be restructured due to cache format changes